+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
  - Vump badge
+ - Bower module
 
 ## [2.0.0] - 2019-03-14
 ### Added